Audio Transcribing/requirements.txt
google-cloud-speech
google-api-python-client
tqdm